--- a/package.json
+++ b/package.json
@@ -24,6 +24,7 @@
         "@rollup/plugin-node-resolve": "^7.1.3",
         "@rollup/plugin-replace": "^2.3.4",
         "@tailwindcss/forms": "^0.2.1",
+        "@tailwindcss/jit": "^0.1.1",
         "alpinejs": "^2.8.1",
         "cypress": "^6.6.0",
         "cypress-iframe": "^1.0.1",