--- a/packages/shell-chrome/views/panel.edge
+++ b/packages/shell-chrome/views/panel.edge
@@ -42,5 +42,6 @@
             </div>
         </div>
         <script src="./panel.js"></script>
+        <script src="./alpine.js"></script>
     </body>
 </html>