--- a/packages/simulator/index.html
+++ b/packages/simulator/index.html
@@ -16,5 +16,6 @@
         <!-- see ./dev.js for injection of panel.html/setup -->
         <div id="devtools-container" class="h-half"></div>
         <script src="./dev.js"></script>
+        <script src="./alpine.js"></script>
     </body>
 </html>