--- a/postcss.config.js
+++ b/postcss.config.js
@@ -1,3 +1,5 @@
 module.exports = {
-    plugins: [require('tailwindcss')],
+    plugins: {
+        '@tailwindcss/jit': {},
+    },
 }