… cypress/integration/component-data.js