- add tailwind-jit (https://github.com/tailwindlabs/tailwindcss-jit)
  - Outcome:
    - uncached: https://github.com/alpine-collective/alpinejs-devtools/pull/185/checks?check_run_id=2129099792 (~2m, similar to cached run before these changes), we save 4s on "build" and ~20s on "npm run start:ci"
    - cached: https://github.com/alpine-collective/alpinejs-devtools/pull/185/checks?check_run_id=2129115704 (~1m30, 20-30s faster than before these changes), we save 4s on "build" and ~15s on "npm run start:ci"
- don't bundle Alpine.js into `panel.js`, copy + include it using `<script>` in HTML
  - allows us to switch to IIFE output for `panel.js`
  - outcome: https://github.com/alpine-collective/alpinejs-devtools/pull/185/checks?check_run_id=2129307322
  - build is slightly faster (not noticeably on CI)
  - `start:ci`/Cypress run is a touch faster (not noticeably on CI)
  - this change will mainly be noticeable locally on rebuild when `panel.js` changes
  - artifact is slightly bigger (due to Alpine not being minified + IIFE output)
- split component "data" and component "list" tests